+++ NEWS.md
@@ -4,6 +4,9 @@
 - Add param `taxa_names_from_physeq` (default FALSE) to `subset_taxa_pq()` 
 - Add param `rarefy_by_sample` (default FALSE) to function `ggbetween_pq()`
 - Add function `are_modality_even_depth()` to test if samples depth significantly vary among the modalities of a factor
+- Add functions `merge_taxa_vec()` and `merge_samples2()` from the [speedyseq](https://github.com/mikemc/speedyseq/) package into MiscMetabar to decrease package dependencies (Thanks to Mike R. Mclaren)
+- Add function `reorder_taxa_pq()` in order to replace the unique call to package MicroViz to decrease package dependencies.
+- Add functions `get_funguild_db()` and `funguild_assign()` from the [FUNGuildR](https://github.com/brendanf/FUNGuildR/) package into MiscMetabar to decrease package dependencies 
 
 # MiscMetabar 0.52